pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "substack-mcp-plus"
version = "1.0.0"
description = "Unofficial MCP server for Substack with rich text formatting (not affiliated with Substack Inc.)"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "ty13r"}
]
keywords = ["mcp", "substack", "api", "automation", "publishing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-substack>=0.1.15",
    "mcp>=1.10.0",
    "markdown2>=2.4.10",
    "pydantic>=2.5.0",
    "python-dotenv>=0.21.0,<1.0.0",
    "aiohttp>=3.9.0",
    "beautifulsoup4>=4.12.0",
    "playwright>=1.40.0",
    "cryptography>=41.0.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

# Note: This package is distributed via NPM only
# PyPI distribution has been removed in favor of NPM for easier installation

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312"]
include = '\.pyi?$'

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = ["E203", "E266", "E501", "B904", "E722"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true
strict_optional = true

# Exclude directories with type errors for now
exclude = [
    "src/tools/",
    "src/converters/",
    "src/utils/",
    "src/server.py",
    "src/server_mcp.py",
]

# Ignore missing type stubs for third-party libraries
[[tool.mypy.overrides]]
module = [
    "substack",
    "substack.*",
    "mcp.*",
    "markdown2",
    "playwright.*",
    "keyring",
    "keyring.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
    "requires_auth: marks tests that require Substack authentication",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]